resolve conflicts after rebase, format files, clean design
