+++ package.json
@@ -35,6 +35,7 @@
     "react-hook-form": "^7.62.0",
     "tailwind-merge": "^3.3.1",
     "tailwindcss": "^4.1.13",
+    "uuid": "^13.0.0",
     "zod": "^4.1.9",
     "uuid": "^13.0.0",
     "zustand": "^5.0.8"
