--- src/dateHelpers.ts
@@ -1,11 +1,0 @@
-import {formatDate, isSameDay, isSameMonth, isSameWeek} from "date-fns";
-
-export function  dayDisplay(start:Date,end:Date){
-    if(isSameDay(start, end)){
-        return formatDate(start, "h:mm a") +' -  ' + formatDate(end, " h:mm a ");
-    }
-    else if(isSameWeek(start,end) || isSameMonth(start,end))
-        return formatDate(start, "h:mm a")+'-'+formatDate(end, " EEE, LLL d h:mm a ");
-    else
-        return formatDate(start, "h:mm a")+'-'+formatDate(end, " LLL yyyy d h:mm a ");
-}
